\name{Sim.DiffGUI}
\alias{Sim.DiffGUI}
\title{
GUI for Sim.DiffProc Package
}
\usage{
Sim.DiffGUI()
}
\description{
A Graphical User Interface for some functions in the Sim.DiffProc package
}
\author{
Guidoum Arsalane, Boukhetala Kamal.
}
\examples{
\donttest{ Sim.DiffGUI()}
}
\keyword{ Diffusion Process }
\keyword{ Environment R }
\keyword{ Simulation }
\keyword{ Financial models }
\keyword{ Stochastic Differential Equation }
\keyword{Numerical Solution of Stochastic Differential Equation}
\keyword{Graphical User Interface GUI}

