\name{Sim.DiffProcGUI-package}
\alias{Sim.DiffProcGUI-package}
\alias{Sim.DiffProcGUI}
\docType{package}
\title{
Graphical User Interface for Simulation of Diffusion Processes 
}
\description{
A Graphical User Interface for some functions in the Sim.DiffProc package
}
\details{
\tabular{ll}{
Package: \tab Sim.DiffProcGUI\cr
Type: \tab Package\cr
Version: \tab 2.5\cr
Date: \tab 2012-06-11\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Boukhetala Kamal <kboukhetala@usthb.dz>, Guidoum Arsalane <starsalane@gmail.com>.

Maintainer: Guidoum Arsalane <starsalane@gmail.com>
}
\references{
\enumerate{
\item Dalgaard P. A Primer on the R-Tcl/Tk Package. R News, 1(3), pp. 27-31, 2001.
\item Dalgaard P. Changes to the R-Tcl/Tk Package. R News, 2(3), pp. 25-71, 2002.
\item Welch B. Practical Programming in Tcl and Tk. 3nd edition. Prentice Hall PTR Upper Saddle River, NJ, USA. ISBN 0-13-022028-0, 2000.
\item S. Baillargeon. Programmation en R: incorporation de code C et creation de packages, 2006.
\item R Development Core Team, 2012. Writing R extensions, ISBN 3-900051-07-0, version 2.15.0 (2012-03-30).
}
}
\keyword{Software Engineering}
\keyword{Graphical User Interface GUI}
\keyword{ Diffusion Process Multidimensional}
\keyword{ Diffusion Process }
\keyword{ Environment R }
\keyword{ Simulation }
\keyword{ Actuarial Modeling }
\keyword{ Financial Models }
\keyword{ Stochastic Differential Equation }
\keyword{ Solution of SDE One-Dimensional}
\keyword{ Solution of SDE Two-Dimensional}
\keyword{ Solution of SDE Three-Dimensional}
\keyword{ Attractive Model}
\keyword{ Statistical Analysis }
\examples{
\donttest{
 Sim.DiffGUI()}
}

