\name{Sim.DiffGUI}
\alias{Sim.DiffGUI}
\title{
GUI for Sim.DiffProc
}
\usage{
Sim.DiffGUI()
}
\description{
A Graphical User Interface for some functions in the Sim.DiffProc package
}
\author{
BOUKHETALA Kamal ,GUIDOUM Arsalane.
}
\examples{
Sim.DiffGUI()
}
\keyword{ Diffusion Process }
\keyword{ Environment R }
\keyword{ Simulation }
\keyword{ financial models }
\keyword{ Stochastic Differential Equation }
\keyword{Numerical Solution of Stochastic Differential Equation}
