\name{Sim.DiffProcGUI-package}
\alias{Sim.DiffProcGUI-package}
\alias{Sim.DiffProcGUI}
\docType{package}
\title{
Graphical User Interface for Simulation of Diffusion Processes 
}
\description{
A Graphical User Interface for some functions in the Sim.DiffProc package
}
\details{
\tabular{ll}{
Package: \tab Sim.DiffProcGUI\cr
Type: \tab Package\cr
Version: \tab 2.0\cr
Date: \tab 2011-02-26\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
BOUKHETALA Kamal <kboukhetala@usthb.dz>, GUIDOUM Arsalane <starsalane@gmail.com>. 

Maintainer: BOUKHETALA Kamal <kboukhetala@usthb.dz>
}
\references{
\enumerate{\item Franck Jedrzejewski. Modeles aleatoires et physique probabiliste, Springer, 2009.
\item K.Boukhetala, Estimation of the first passage time distribution for a simulated diffusion process, Maghreb Math.Rev, Vol.7, No 1, Jun 1998, pp. 1-25. \item K.Boukhetala, Simulation study of a dispersion about an attractive centre. In proceedings of 11th Symposium Computational Statistics, edited by R.Dutter and W.Grossman, Wien , Austria, 1994, pp. 128-130. \item K.Boukhetala,Modelling and simulation of a dispersion pollutant with attractive centre, Edited by Computational Mechanics Publications, Southampton ,U.K and Computational Mechanics Inc, Boston, USA, pp. 245-252. \item K.Boukhetala, Kernel density of the exit time in a simulated diffusion, les Annales Maghrebines De L ingenieur, Vol , 12, N Hors Serie. Novembre 1998, Tome II, pp 587-589. \item T. Rolski, H. Schmidli, V. Schmidt and J. Teugels, Stochastic Processes for Insurance and Finance,John Wiley & Sons, 1998.
\item Fima C Klebaner. Introduction to stochastic calculus with application (Second Edition), Imperial College Press (ICP), 2005.
\item LAWRENCE C.EVANS. An introduction to stochastic differential equations (Version 1.2), Department of Mathematics (UC BERKELEY).
\item Hui-Hsiung Kuo. Introduction to stochastic integration, Springer, 2006.
\item E.Allen. Modeling with Ito stochastic differential equations, Springer, 2007.
\item Peter E.Kloeden, Eckhard Platen, Numerical solution of stochastic differential equations, Springer, 1995.
\item Douglas Henderson, Peter Plaschko, Stochastic differential equations in science and engineering,World Scientific, 2006.
\item A.Greiner, W.Strittmatter,and J.Honerkamp, Numerical Integration of Stochastic Differential Equations, Journal of Statistical Physics, Vol. 51, Nos. 1/2, 1988.
\item YOSHIHIRO SAITO, TAKETOMO MITSUI, SIMULATION OF STOCHASTIC DIFFERENTIAL EQUATIONS, Ann.Inst.Statist.Math, Vol. 45, No.3,419-432 (1993).
\item FRANCOIS-ERIC RACICOT, RAYMOND THEORET, Finance computationnelle et gestion des risques, Presses de universite du Quebec, 2006.
\item Avner Friedman, Stochastic differential equations and applications, Volume 1, ACADEMIC PRESS, 1975.
}
}
\keyword{ Attractive Model}
\keyword{ Diffusion Process Multidimensional}
\keyword{ Diffusion Process }
\keyword{ Environment R }
\keyword{ Simulation }
\keyword{ Actuarial Modeling }
\keyword{ Financial Models }
\keyword{ Stochastic Differential Equation }
\keyword{ Solution of SDE One-Dimensional}
\keyword{ Solution of SDE Two-Dimensional}
\keyword{ Statistical Analysis }
\examples{
Sim.DiffGUI()
}

