\name{Sim.DiffProcGUI-package}
\alias{Sim.DiffProcGUI-package}
\alias{Sim.DiffProcGUI}
\docType{package}
\title{
Graphical User Interface for Simulation of Diffusion Processes 
}
\description{
A Graphical User Interface for some functions in the Sim.DiffProc package
}
\details{
\tabular{ll}{
Package: \tab Sim.DiffProcGUI\cr
Type: \tab Package\cr
Version: \tab 2.2\cr
Date: \tab 2012-02-14\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
BOUKHETALA Kamal <kboukhetala@usthb.dz>, GUIDOUM Arsalane <starsalane@gmail.com>. 

Maintainer: BOUKHETALA Kamal <kboukhetala@usthb.dz>
}
\references{
\enumerate{
\item Dalgaard P. A Primer on the R-Tcl/Tk Package. R News, 1(3), pp. 27-31, 2001.
\item Dalgaard P. Changes to the R-Tcl/Tk Package. R News, 2(3), pp. 25-71, 2002.
\item Welch B. Practical Programming in Tcl and Tk. 3nd edition. Prentice Hall PTR Upper Saddle River, NJ, USA. ISBN 0-13-022028-0, 2000.
\item S. Baillargeon. Programmation en R: incorporation de code C et creation de packages, 2006.
\item R Development Core Team, 2011. Writing R extensions, ISBN 3-900051-11-9, version 2.13.1 (2011-07-08).
}
}
\keyword{Software Engineering}
\keyword{ Attractive Model}
\keyword{ Diffusion Process Multidimensional}
\keyword{ Diffusion Process }
\keyword{ Environment R }
\keyword{ Simulation }
\keyword{ Actuarial Modeling }
\keyword{ Financial Models }
\keyword{ Stochastic Differential Equation }
\keyword{ Solution of SDE One-Dimensional}
\keyword{ Solution of SDE Two-Dimensional}
\keyword{ Solution of SDE Three-Dimensional}
\keyword{ Statistical Analysis }
\examples{
## Sim.DiffGUI()
}

